+++ cmd/main.go
@@ -210,6 +210,13 @@ func main() {
 		Client: mgr.GetClient(),
 		Scheme: mgr.GetScheme(),
 	}
+	if err := (&controller.PrometheusClusterSetReconciler{
+		Client: mgr.GetClient(),
+		Scheme: mgr.GetScheme(),
+	}).SetupWithManager(mgr); err != nil {
+		setupLog.Error(err, "Failed to create controller", "controller", "PrometheusClusterSet")
+		os.Exit(1)
+	}
 	// +kubebuilder:scaffold:builder
 
 	signalCtx := ctrl.SetupSignalHandler()